sse_avx: fixes, vec mul test

+++ services/sse_avx/compiler/frontend.h
@@ -127,7 +127,7 @@ struct Instruction
         OperandType type;
         union
         {
-            uint32_t imm;
+            uint64_t imm;
             Register reg;
             const char* labelId;
         };
